+++ lib/util/util_id.c
@@ -26,7 +26,7 @@
 ****************************************************************************/
 
 bool add_gid_to_array_unique(TALLOC_CTX *mem_ctx, gid_t gid,
-			     gid_t **gids, size_t *num_gids)
+			     gid_t **gids, uint32_t *num_gids)
 {
 	int i;
 
@@ -59,7 +59,7 @@ bool add_gid_to_array_unique(TALLOC_CTX *mem_ctx, gid_t gid,
 ****************************************************************************/
 
 bool add_uid_to_array_unique(TALLOC_CTX *mem_ctx, uid_t uid,
-			     uid_t **uids, size_t *num_uids)
+			     uid_t **uids, uint32_t *num_uids)
 {
 	int i;
